refactored the itinerary list and itin preview to display all saved itineraries on page load.
