+++ index.html
@@ -46,7 +46,7 @@
                             </article>
                             <article class="container--eatery">
                             </article>
-                            <button class="save-btn disabled" id="saveItinerary" disabled>
+                            <button class="save-btn btn--save--itin disabled" id="saveItinerary" disabled>
                                 Save Itinerary
                             </button>
                         </section>
