+++ scripts/itinerary-preview/itineraryHTML.js
@@ -1,14 +1,10 @@
-export const itineraryHTML = (itenObj) => {
-    console.log(itenObj)
+export const itineraryHTML = (itinObj) => {
     return `
-    <br>
-    <fieldset>
-          <section>
-                <div><b>Park: </b>${itenObj.park}</div>
-                <div><b>Attraction: </b>${itenObj.bizarrie}</div>
-                <div><b>Restaurant: </b>${itenObj.eatery}</div>
+          <section class="itin-card">
+                <div class="itin-card--park">Park: ${ itinObj.park }</div>
+                <div class="itin-card--biz">Attraction: ${ itinObj.bizarrie }</div>
+                <div class="itin-card--eat">Restaurant: ${ itinObj.eatery }</div>
+                <div class="itin-card--date">Date Created: ${ new Date(itinObj.timestamp).toLocaleDateString('en-US') }
           </section>
-      </fieldset>
-      <br>
     `
 }
