+++ styles/main.css
@@ -11,8 +11,11 @@
     color: #b2b044; 
     */
 
+/* * {
+    border: 1px solid hotpink;
+} */
+
 body {
-    width: 100%;
     font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
     background-color: white;
     color: #191919;
@@ -34,6 +37,8 @@ h3 {
     height: auto;
     flex-direction: column;
     margin: 0;
+    position: fixed;
+    overflow: hidden;
 }
 
 .container--content {
@@ -53,4 +58,14 @@ h3 {
 .container--right {
     display: flex;
     width: 20%;
+}
+
+::-webkit-scrollbar {
+    -webkit-appearance: none;
+    width: 8px;
+}
+
+::-webkit-scrollbar-thumb {
+    border: 2px solid #5b7c8e;
+    background-color: white;
 }
