+++ styles/select.css
@@ -28,4 +28,5 @@
 
 .dropdown {
     margin: 1rem;
+    width: 90%;
 }
